# swaggerAPI.yml
openapi: "3.0.2"
info:
  title: Task Manager API
  description: "A very simple task manager API used for learning purposes"
  version: "1.0"
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A **unique** JWT token is generated upon registration and user login. It must be used in the request header called Authorization. Eg. ```Authorization:Bearer {{token}}```
  schemas:
...
    RequestCreateTask:
      type: object
      required:
        - author
        - description
      properties:
        description:
          type: string
          description: The description of the task
        completed:
          type: boolean
          description: The status of the task. (DEFAULT:false)
        author:
          type: "mongoose.Schema.Types.ObjectId"
          description: The user's id who owns the task. (Auto inserted, thus no need to explicitly insert)
...
paths:
...
  /tasks:
    get:
      security:
        - BearerAuth: []
      tags:
        - Task
      summary: Get all tasks
      parameters:
        - name: completed
          in: query
          description: True or False for the tasks
          schema:
            type: boolean
        - name: page
          in: query
          description: The page number to access
          schema:
            type: integer
        - name: limit
          in: query
          description: The number of a page
          schema:
            type: integer
        - name: sortBy
          in: query
          description: The property of the task to be sorted either by desc or asc ```sortBy=completed:desc```
          schema:
            type: string
      responses:
        200:
          description: Successfully obtained current user's task/s
          content:
            application/json:
              schema:
                type: array
                items:
                  example:
                    _id: 611215dc105743549012dc31
                    description: Help a british agent 272 escape from East Berlin
                    completed: false
                    author: 61121010d676405690cee467
                    createdAt: 2021-08-10T02:17:03.009Z
                    updatedAt: 2021-08-10T05:03:30.862Z
        401:
          description: Unauthorized Access
          content:
            application/json:
              example:
                error: Unauthorized Access
    post:
      security:
        - BearerAuth: []
      tags:
        - Task
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: Task
              $ref: "#/components/schemas/RequestCreateTask"
            example:
              description: Second mission to Thailand

      responses:
        200:
          description: Successfully created a new task
          content:
            application/json:
              example:
                _id: 61121b9652352c4b505aad60
                description: Second mission to Thailand
                completed: false
                author: 61121010d676405690cee467
                createdAt: 2021-08-10T06:24:22.660Z
                updatedAt: 2021-08-10T06:24:22.660Z
        401:
          description: Unauthorized Access
          content:
            application/json:
              example:
                error: Unauthorized Access
        400:
          description: Description cannot be empty
          content:
            application/json:
              example:
                error: Description is required
...